# ViewportsLocater-LucasGarcia-Back/VIEWPORTS_LOCATER/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Viewports Locater API",
        "description": "Documentacion de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/miradores/{id}": {
            "get": {
                "tags": [
                    "Miradores"
                ],
                "summary": "Obtener un mirador por ID",
                "operationId": "showMirador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del mirador con provincia, rutas y fotos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mirador": {
                                            "$ref": "#/components/schemas/Mirador"
                                        },
                                        "provincia": {
                                            "$ref": "#/components/schemas/Provincia"
                                        },
                                        "rutas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ruta"
                                            }
                                        },
                                        "fotos": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mirador no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Miradores"
                ],
                "summary": "Actualizar un mirador existente",
                "operationId": "updateMirador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Mirador del Tibidabo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción actualizada",
                                        "nullable": true
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 41.4236
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.1197
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mirador actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mirador": {
                                            "$ref": "#/components/schemas/Mirador"
                                        },
                                        "provincia": {
                                            "$ref": "#/components/schemas/Provincia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mirador no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Miradores"
                ],
                "summary": "Eliminar un mirador",
                "operationId": "destroyMirador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mirador eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mirador eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mirador no encontrado"
                    }
                }
            }
        },
...
        "/api/rutas/{id}": {
            "get": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Obtener una ruta por ID",
                "operationId": "showRuta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la ruta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ruta": {
                                            "$ref": "#/components/schemas/Ruta"
                                        },
                                        "mirador": {
                                            "$ref": "#/components/schemas/Mirador"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Actualizar una ruta existente",
                "operationId": "updateRuta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Ruta del Pico Mayor"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción actualizada",
                                        "nullable": true
                                    },
                                    "distancia_km": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15,
                                        "nullable": true
                                    },
                                    "duracion_estimada_min": {
                                        "type": "integer",
                                        "example": 200,
                                        "nullable": true
                                    },
                                    "dificultad": {
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Alta",
                                        "nullable": true
                                    },
                                    "enlace_maps": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "https://maps.google.com/...",
                                        "nullable": true
                                    },
                                    "gpx_url": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "https://ejemplo.com/ruta-nueva.gpx",
                                        "nullable": true
                                    },
                                    "mirador_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ruta actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ruta": {
                                            "$ref": "#/components/schemas/Ruta"
                                        },
                                        "mirador": {
                                            "$ref": "#/components/schemas/Mirador"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Eliminar una ruta",
                "operationId": "destroyRuta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruta eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ruta eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruta no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mirador": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Mirador del Tibidabo"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Vistas panorámicas de Barcelona",
                        "nullable": true
                    },
                    "latitud": {
                        "type": "number",
                        "format": "float",
                        "example": 41.4236
                    },
                    "longitud": {
                        "type": "number",
                        "format": "float",
                        "example": 2.1197
                    },
                    "provincia_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Provincia": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Barcelona"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Ruta": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Ruta del Pico Mayor"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Ruta de montaña con vistas panorámicas",
                        "nullable": true
                    },
                    "distancia_km": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5,
                        "nullable": true
                    },
                    "duracion_estimada_min": {
                        "type": "integer",
                        "example": 180,
                        "nullable": true
                    },
                    "dificultad": {
                        "type": "string",
                        "example": "Media",
                        "nullable": true
                    },
                    "enlace_maps": {
                        "type": "string",
                        "example": "https://maps.google.com/...",
                        "nullable": true
                    },
                    "gpx_url": {
                        "type": "string",
                        "example": "https://ejemplo.com/ruta.gpx",
                        "nullable": true
                    },
                    "mirador_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}